pyproject.toml
[project]
name = "sciencedirect-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "pydantic-ai-slim[openai]>=0.8.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "typer>=0.17.3",
]
